requirements.txt
attrs==21.4.0
bleach==4.1.0
build==0.7.0
bump2version==1.0.1
bumpversion==0.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
docutils==0.18.1
flake8==4.0.1
idna==3.3
importlib-metadata==4.10.1
iniconfig==1.1.1
keyring==23.5.0
mccabe==0.6.1
packaging==21.3
pep517==0.12.0
pkginfo==1.8.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
readme-renderer==32.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
six==1.16.0
toml==0.10.2
tomli==2.0.0
tqdm==4.62.3
twine==3.7.1
urllib3==1.26.8
webencodings==0.5.1
websocket-client==1.2.3
websockets==10.1
zipp==3.7.0
